rename param.hh/cc to hist_param.hh/cc to fix xcode build
